Update email and phone max length, and update GetTotalRevenue function 'better performance'.

+++ RestaurantReservation.Db/Entities/Order.cs
@@ -8,6 +8,6 @@ public class Order
     [Required] public int ReservationID { get; set; }
     [Required] public int EmployeeID { get; set; }
     [Required] public DateTime OrderDate { get; set; } = DateTime.Now;
-    [Required] [Range(0.01, 10000)] public double TotalAmount { get; set; }
+    [Required] [Range(0.01, 10000)] public decimal TotalAmount { get; set; }
     public List<OrderItem> OrderItems { get; set; }
 }
